+++ lib/group/digest2/formatData.js
