+++ lib/group/digest2/index.js
