+++ lib/group/digest2/personalizeData.js
