+++ lib/group/digest2/util.js
